# docs/api/viewlayouts-openapi.yaml
openapi: 3.0.3
info:
  title: ViewLayouts API
  description: REST Level 3 API for managing visual layout containers and element positions
  version: 1.0.0

servers:
  - url: /api/v1
    description: API v1

tags:
  - name: Layouts
    description: Layout container operations
  - name: Elements
    description: Element position operations

paths:
  /layouts/{contextType}/{contextRef}:
    parameters:
      - name: contextType
        in: path
        required: true
        schema:
          type: string
          enum:
            - architecture-canvas
            - business-domain-grid
        description: Type of UI context
      - name: contextRef
        in: path
        required: true
        schema:
          type: string
        description: Reference ID (view_id or domain_id)

    get:
      summary: Get layout container
      description: Retrieve layout container with all element positions
      operationId: getLayout
      tags:
        - Layouts
      responses:
        '200':
          description: Layout found
          headers:
            ETag:
              schema:
                type: string
              description: Version number for optimistic locking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayoutContainer'
        '404':
          description: Layout not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundWithCreateLink'

    put:
      summary: Create or update layout container
      description: Create a new layout container or update preferences (upsert)
      operationId: upsertLayout
      tags:
        - Layouts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertLayoutRequest'
      responses:
        '200':
          description: Layout updated
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayoutContainer'
        '201':
          description: Layout created
          headers:
            Location:
              schema:
                type: string
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayoutContainer'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    delete:
      summary: Delete layout container
      description: Delete layout container and all element positions
      operationId: deleteLayout
      tags:
        - Layouts
      responses:
        '204':
          description: Layout deleted
        '404':
          description: Layout not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /layouts/{contextType}/{contextRef}/preferences:
    parameters:
      - name: contextType
        in: path
        required: true
        schema:
          type: string
          enum:
            - architecture-canvas
            - business-domain-grid
      - name: contextRef
        in: path
        required: true
        schema:
          type: string

    patch:
      summary: Update preferences
      description: Update layout preferences with optimistic locking
      operationId: updateLayoutPreferences
      tags:
        - Layouts
      parameters:
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
          description: ETag for optimistic locking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
      responses:
        '200':
          description: Preferences updated
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayoutContainerSummary'
        '404':
          description: Layout not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '412':
          description: Version conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'

  /layouts/{contextType}/{contextRef}/elements:
    parameters:
      - name: contextType
        in: path
        required: true
        schema:
          type: string
          enum:
            - architecture-canvas
            - business-domain-grid
      - name: contextRef
        in: path
        required: true
        schema:
          type: string

    patch:
      summary: Batch update element positions
      description: Update multiple element positions atomically
      operationId: batchUpdateElements
      tags:
        - Elements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchUpdateRequest'
      responses:
        '200':
          description: All updates succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchUpdateResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Layout not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /layouts/{contextType}/{contextRef}/elements/{elementId}:
    parameters:
      - name: contextType
        in: path
        required: true
        schema:
          type: string
          enum:
            - architecture-canvas
            - business-domain-grid
      - name: contextRef
        in: path
        required: true
        schema:
          type: string
      - name: elementId
        in: path
        required: true
        schema:
          type: string
        description: Element identifier (component_id or capability_id)

    put:
      summary: Upsert element position
      description: Create or replace element position (idempotent, last-write-wins)
      operationId: upsertElementPosition
      tags:
        - Elements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementPositionInput'
      responses:
        '200':
          description: Position updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementPosition'
        '201':
          description: Position created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementPosition'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Layout not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete element position
      description: Remove element position from layout
      operationId: deleteElementPosition
      tags:
        - Elements
      responses:
        '204':
          description: Position deleted
        '404':
          description: Layout or element not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    LayoutContainer:
      type: object
      required:
        - id
        - contextType
        - contextRef
        - preferences
        - elements
        - version
        - createdAt
        - updatedAt
        - _links
      properties:
        id:
          type: string
          format: uuid
        contextType:
          type: string
          enum:
            - architecture-canvas
            - business-domain-grid
        contextRef:
          type: string
        preferences:
          type: object
          additionalProperties: true
        elements:
          type: array
          items:
            $ref: '#/components/schemas/ElementPosition'
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        _links:
          $ref: '#/components/schemas/LayoutLinks'

    LayoutContainerSummary:
      type: object
      required:
        - id
        - contextType
        - contextRef
        - preferences
        - version
        - _links
      properties:
        id:
          type: string
          format: uuid
        contextType:
          type: string
        contextRef:
          type: string
        preferences:
          type: object
          additionalProperties: true
        version:
          type: integer
        _links:
          $ref: '#/components/schemas/SelfLink'

    ElementPosition:
      type: object
      required:
        - elementId
        - x
        - y
        - _links
      properties:
        elementId:
          type: string
        x:
          type: number
          format: double
        y:
          type: number
          format: double
        width:
          type: number
          format: double
        height:
          type: number
          format: double
        customColor:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        sortOrder:
          type: integer
        _links:
          $ref: '#/components/schemas/ElementLinks'

    UpsertLayoutRequest:
      type: object
      properties:
        preferences:
          type: object
          additionalProperties: true

    UpdatePreferencesRequest:
      type: object
      additionalProperties: true

    ElementPositionInput:
      type: object
      required:
        - x
        - y
      properties:
        x:
          type: number
          format: double
        y:
          type: number
          format: double
        width:
          type: number
          format: double
        height:
          type: number
          format: double
        customColor:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        sortOrder:
          type: integer

    BatchUpdateRequest:
      type: object
      required:
        - updates
      properties:
        updates:
          type: array
          items:
            allOf:
              - type: object
                required:
                  - elementId
                properties:
                  elementId:
                    type: string
              - $ref: '#/components/schemas/ElementPositionInput'

    BatchUpdateResponse:
      type: object
      required:
        - updated
        - elements
        - _links
      properties:
        updated:
          type: integer
        elements:
          type: array
          items:
            $ref: '#/components/schemas/ElementPosition'
        _links:
          type: object
          properties:
            self:
              $ref: '#/components/schemas/Link'
            layout:
              $ref: '#/components/schemas/Link'

    LayoutLinks:
      type: object
      properties:
        self:
          $ref: '#/components/schemas/Link'
        updatePreferences:
          $ref: '#/components/schemas/Link'
        batchUpdate:
          $ref: '#/components/schemas/Link'
        delete:
          $ref: '#/components/schemas/Link'

    ElementLinks:
      type: object
      properties:
        self:
          $ref: '#/components/schemas/Link'
        layout:
          $ref: '#/components/schemas/Link'
        update:
          $ref: '#/components/schemas/Link'
        delete:
          $ref: '#/components/schemas/Link'

    SelfLink:
      type: object
      properties:
        self:
          $ref: '#/components/schemas/Link'

    Link:
      type: object
      required:
        - href
      properties:
        href:
          type: string
          format: uri-reference
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string

    NotFoundWithCreateLink:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            _links:
              type: object
              properties:
                create:
                  $ref: '#/components/schemas/Link'

    ConflictError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            currentVersion:
              type: integer
            _links:
              type: object
              properties:
                current:
                  $ref: '#/components/schemas/Link'

    ValidationError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              type: object
              additionalProperties:
                type: string

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: []